+++ app/(tabs)/settings.tsx
@@ -1,5 +1,6 @@
 import { Stack } from 'expo-router';
 import { View, Text, StyleSheet } from 'react-native';
+import React from 'react';
 
 export default function Tab() {
   return (
